Fix bug in waitfor and expose command
