+++ cmd/get/component/cmd.go
